pytest: Separate patient tests
